Refactor modals: move to components/More and rename to Modal variants
